chore(releasewf): Fix bump_cdk.sh tagging logic.
bump_cdk.sh was adding a tag while not commiting the changes due
to which the previous commit was wrongly tagged with the new version.
This change will make sure script is ran from clean main branch and
will create a commit from the changes needed to bump CDK version.

It will provide instructions on how to push them but that is still up
to the user who can first do further verification.