chore(releasewf): Extend bump_cdk.sh actions
bump_cdk.sh bumps the version and now also will perform a compile
run which runs the local tests and compiles the components and doc.

Additionally it will output the git status such that a user can
validate changes that have been performed by the compile run. For
example changes in aws-solutions-constructs which result in different
test snapshots. This allows for easy review prior to pushing to remote.