feat(deps) upgrade CDK to 1.102.0
As part of this upgrade also change the code to only pass in existing
objects if nor properties are specified. This is required since
aws-solutions-constructs 1.102.0 has a breaking change where if you
send an existing resource and props for a new resource, the construct
 will now throw an error. This code change makes sure that situation
does not arise. Also updated tests to go through branch paths.